Comments and minor adjustments
I added comments and made some minor adjustments: A simple style sheet  as well as an array and counter for correct answers in addition to incorrect ones. This version is supposed to be presented in the seminar as starting point, i.e. the simplest vocabulary trainer possible. Next goal: Randomized order & repeated requests.