Version 2: Random order and repeated requests
I added an ordering array and a shuffle function in order to achieve a random question order. Furthermore, the user is asked each word repeatedly until s/he either translates it correctly or types "Stop" (as a way out). To achieve this, the general counter is only set +1 if these conditions are met.