for loops and find correct answer
I replaced all while-loops with for-loops. Incorrect answers are not asked immediately in the first session but in a second session.
Lines 92-109: Tried to find a way to track the correct answer to previously incorrectly answered  question in original array. Worked, but I need a new variable (Neu) to save correct answer. That is problematic because I have to find it later on (line 192 instead of Falsch2), so I will go back to my older solution in the next version and use Falsch2 instead.